auth/requirements.txt
bcrypt
flask-security 
flask-sqlalchemy
